Added a confirmation message for the clear log by user agents in admin's result page

+++ static/js/admin.js
@@ -28,28 +28,30 @@ $(function() {
     $('#clearLogUserAgents').click(function( event ) {
 		event.preventDefault();
 
-    	var userAgents = $('#selUserAgent').val();
-    	var clearUserAgentsUrl = $(this).attr('url');
+		if (confirm("Do you really want to clear the log for these user agents?")) {
+	    	var userAgents = $('#selUserAgent').val();
+	    	var clearUserAgentsUrl = $(this).attr('url');
 
-		if (userAgents != null) {
-	    	var userAgentsString = '';
-	    	for (var i = userAgents.length - 1; i >= 0; i--) {
-	    		userAgentsString = 'selUserAgent=' + encodeURIComponent(userAgents[i]) + '&' + userAgentsString;
-	    	};
-	    	userAgentsString = userAgentsString.substring(0, userAgentsString.length - 1);
-	    	clearUserAgentsUrl = clearUserAgentsUrl + '?' + userAgentsString;
+			if (userAgents != null) {
+		    	var userAgentsString = '';
+		    	for (var i = userAgents.length - 1; i >= 0; i--) {
+		    		userAgentsString = 'selUserAgent=' + encodeURIComponent(userAgents[i]) + '&' + userAgentsString;
+		    	};
+		    	userAgentsString = userAgentsString.substring(0, userAgentsString.length - 1);
+		    	clearUserAgentsUrl = clearUserAgentsUrl + '?' + userAgentsString;
 
-	    	$.ajax({
-	          type: "DELETE",
-	          url: clearUserAgentsUrl,
-	          success: function() {
-	            alert('Success');
-	          },
-	          error: function() {
-	            alert('Error')
-	          }
-	        });
-	    }
+		    	$.ajax({
+		          type: "DELETE",
+		          url: clearUserAgentsUrl,
+		          success: function() {
+		            alert('Success');
+		          },
+		          error: function() {
+		            alert('Error')
+		          }
+		        });
+		    }
+		}
     });
 
     $('#frmUserAgent').submit(function( event ) {
